nutrifit_backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
google-generativeai==0.3.2
Pillow==10.1.0
